Remove deprecated/unavailable cask brew command
As of Brew 2.6, cask is deprecated (https://brew.sh/2020/12/01/homebrew-2.6.0/), and it seems to have been removed from the latest versions of brew, so this command can fail on new/updated versions of brew. I think `docker` should be what we want, but if we still want the cask, we can install that instead.